added domain types, can filter download queue for website home pages only, and fixed issues with duplicate downloads & articles

+++ Sql/Schema/Stored Procedures/Downloads/Download_Delete.sql
@@ -2,7 +2,7 @@
 	DROP PROCEDURE [dbo].[Download_Delete]
 GO
 CREATE PROCEDURE [dbo].[Download_Delete]
-	@qid int = 0
+	@qid bigint = 0
 AS
 	DECLARE @url nvarchar(250)
 	SELECT @url = [url] FROM DownloadQueue WHERE qid=@qid
